step2 gpu: carry the per-marker dosage table in double, not float
The four-entry code->dosage table is the ONLY per-marker input the device gets.
Three of its entries are exact small integers, but the fourth is the imputed
value for missing calls -- 2*altFreq under impute_method: mean -- and narrowing
that to float on the host put a 6e-8 relative error into every missing cell
before the reduction started. That is a difference from the CPU's INPUT, not
from its arithmetic, and no amount of fp64 in the GEMM can undo it.

It costs 32 bytes per marker to carry (against 12,500 bytes of packed
genotypes). In fp32 mode the kernel narrows it itself, so both precision modes
see the same table.

Verified, 8 quantitative traits x 4,950 markers = 39,600 pairs on a copy of the
benchmark .bed with 2.3% missing calls injected and 50 markers pushed over
maxMissRate:

  before   fp64  max |d(-log10 p)| 3.586e-07,  max |dS|/sd(S) 1.210e-05
  after    fp64  max |d(-log10 p)| 0.000e+00,  max |dS|/sd(S) 0.000e+00

i.e. fp64 output with missing genotypes is now byte-identical to the CPU path,
as it already was on data with none. fp32 is unchanged at ~4e-6 / ~1.3e-5.

Also refuse the path (rather than memcpy past the end) if the stacked constants
are not n rows -- the gate already forbids the only way that could happen.

+++ SAIGE_cpp_260716/step2_saige-step2/gpu/gpu_step2.hpp
@@ -90,9 +90,16 @@ void     destroy(Reducer* t_r);
 // Pinned staging buffers owned by the reducer. The caller writes marker data
 // straight into them, so no copy stands between the decode and the H2D.
 //   packed()  t_maxSlots * bpv bytes, bpv = (N+3)/4; slot j at offset j*bpv
-//   lut()     t_maxSlots * 4 floats;  slot j's code->dosage table at 4*j
+//   lut()     t_maxSlots * 4 doubles; slot j's code->dosage table at 4*j.
+//             DOUBLE, not float, in both precision modes: three of its entries
+//             are exact small integers but the fourth is the imputed mean
+//             2*altFreq, and narrowing that on the host would put a 6e-8
+//             relative error into every missing cell before the reduction even
+//             starts -- which is a difference from the CPU's INPUT, not from
+//             its arithmetic. In fp32 mode the kernel narrows it itself, so
+//             the two modes still see the same table.
 unsigned char* packed(Reducer* t_r);
-float*         lut(Reducer* t_r);
+double*        lut(Reducer* t_r);
 std::size_t    bytesPerSlot(const Reducer* t_r);
 
 // Reduce slots [0, t_nSlots). Returns false on any CUDA error, in which case
